Changes in src of the app , changes like logo of app , done , and a basic layout with all types of packages called that will be needed for it , and a basic layout is ready
