+++ README.md
@@ -16,3 +16,5 @@
 - Include the screenshot of the App while submitting your PR.
 
 ## Contributors 
+
+- Manas kashyap
