--- src/app/src/main/res/layout/content_main.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/content_main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="in.asetalias.alias.MainActivity"
-    tools:showIn="@layout/activity_main">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true">
-        <ImageView
-            android:id="@+id/logo"
-            android:layout_width="match_parent"
-            android:layout_height="200dp"
-            android:src="@drawable/alias"
-            android:cropToPadding="false"/>
-        <TextView
-            android:id="@+id/alias_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/alias_full_name"
-            android:layout_below="@id/logo"
-            android:textAppearance="@style/TextAppearance.AppCompat.Headline"
-            android:textAlignment="center"
-            android:textStyle="normal|bold"
-            android:textSize="24sp" />
-    </LinearLayout>
-
-</RelativeLayout>
