+++ src/app/src/main/res/values/strings.xml
@@ -1,5 +1,3 @@
 <resources>
-    <string name="app_name">ALiAS</string>
-    <string name="action_settings">Settings</string>
-    <string name="alias_full_name">Amity Linux Assistance Sapience</string>
+    <string name="app_name">ALIAS</string>
 </resources>
